tests(preprocessing): add new test files
