+++ tests/test_preprocessing_Bus.py
@@ -1,0 +1,3 @@
+def test_buses():
+    from program_files.preprocessing.components import Bus
+    pass
