+++ tests/test_preprocessing_district_heating_clustering.py
@@ -1,0 +1,4 @@
+def test_connect_clustered_dh_to_system():
+    from program_files.preprocessing.components.district_heating_clustering \
+        import connect_clustered_dh_to_system
+    pass
